Implement Redis Adapter configurations

+++ NetSDK/Services/Client/Classes/RedisConfigurationOptions.cs
@@ -11,9 +11,9 @@ namespace Splitio.Services.Client.Classes
         public string Port { get; set; }
         public string Password { get; set; }
         public int? Database { get; set; }
-        public long? ConnectTimeout { get; set; }
-        public long? ConnectRetry { get; set; }
-        public long? SyncTimeout { get; set; }
+        public int? ConnectTimeout { get; set; }
+        public int? ConnectRetry { get; set; }
+        public int? SyncTimeout { get; set; }
         public string UserPrefix { get; set; }
     }
 }
